[#7] Moved ContextProvider from impl to corresponding context logger binding artifacts to minimize dependencies of impl.
