+++ impl/pom.xml
@@ -74,30 +74,6 @@
 			<version>${apache.commonslang.version}</version>
 		</dependency>
 
-		<!-- provided dependencies -->
-        <dependency>
-            <groupId>org.jboss.spec.javax.servlet</groupId>
-            <artifactId>jboss-servlet-api_3.0_spec</artifactId>
-            <version>1.0.0.Final</version>
-            <scope>provided</scope>
-        </dependency>
-
-
-
-        <dependency>
-            <groupId>org.aspectj</groupId>
-            <artifactId>aspectjrt</artifactId>
-            <version>${aspectj.version}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>javax.ejb</groupId>
-            <artifactId>ejb-api</artifactId>
-            <scope>provided</scope>
-        </dependency>
-
-
         <!-- dependecies for unittests -->
         <dependency>
             <groupId>io.tracee.backend</groupId>
@@ -134,6 +110,11 @@
             <groupId>org.reflections</groupId>
             <artifactId>reflections</artifactId>
         </dependency>
+
+		<dependency>
+			<groupId>io.tracee.contextlogger</groupId>
+			<artifactId>testhelper</artifactId>
+		</dependency>
 
     </dependencies>
