+++ impl/src/main/resources/TraceeContextLoggerBasicProfile.properties
@@ -26,47 +26,47 @@ io.tracee.contextlogger.contextprovider.java.JavaThrowableContextProvider.messag
 io.tracee.contextlogger.contextprovider.java.JavaThrowableContextProvider.stacktrace=true
 
 # Watchdog context data
-io.tracee.contextlogger.contextprovider.tracee.WatchdogContextProvider=true
-io.tracee.contextlogger.contextprovider.tracee.WatchdogContextProvider.id=true
-io.tracee.contextlogger.contextprovider.tracee.WatchdogContextProvider.proceedingJoinPoint=true
+io.tracee.contextlogger.watchdog.contextprovider.WatchdogContextProvider=true
+io.tracee.contextlogger.watchdog.contextprovider.WatchdogContextProvider.id=true
+io.tracee.contextlogger.watchdog.contextprovider.WatchdogContextProvider.proceedingJoinPoint=true
 
 # ASPECTJ - ProceedingJoinPoint
-io.tracee.contextlogger.contextprovider.aspectj.AspectjProceedingJoinPointContextProvider=true
-io.tracee.contextlogger.contextprovider.aspectj.AspectjProceedingJoinPointContextProvider.clazz=true
-io.tracee.contextlogger.contextprovider.aspectj.AspectjProceedingJoinPointContextProvider.method=true
-io.tracee.contextlogger.contextprovider.aspectj.AspectjProceedingJoinPointContextProvider.parameters=true
-io.tracee.contextlogger.contextprovider.aspectj.AspectjProceedingJoinPointContextProvider.serializedTargetInstance=false
+io.tracee.contextlogger.watchdog.contextprovider.AspectjProceedingJoinPointContextProvider=true
+io.tracee.contextlogger.watchdog.contextprovider.AspectjProceedingJoinPointContextProvider.clazz=true
+io.tracee.contextlogger.watchdog.contextprovider.AspectjProceedingJoinPointContextProvider.method=true
+io.tracee.contextlogger.watchdog.contextprovider.AspectjProceedingJoinPointContextProvider.parameters=true
+io.tracee.contextlogger.watchdog.contextprovider.AspectjProceedingJoinPointContextProvider.serializedTargetInstance=false
 
 # servlet request
-io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider=true
-io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider.url=true
-io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider.httpMethod=true
-io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider.httpParameters=true
-io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider.httpRequestHeaders=false
-io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider.httpRequestAttributes=false
-io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider.cookies=false
-io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider.httpRemoteAddress=false
-io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider.httpRemoteHost=false
-io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider.httpRemotePort=false
-io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider.scheme=false
-io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider.secure=false
-io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider.contentType=false
-io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider.contentLength=false
-io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider.locale=false
+io.tracee.contextlogger.servlet.contextprovider.ServletRequestContextProvider=true
+io.tracee.contextlogger.servlet.contextprovider.ServletRequestContextProvider.url=true
+io.tracee.contextlogger.servlet.contextprovider.ServletRequestContextProvider.httpMethod=true
+io.tracee.contextlogger.servlet.contextprovider.ServletRequestContextProvider.httpParameters=true
+io.tracee.contextlogger.servlet.contextprovider.ServletRequestContextProvider.httpRequestHeaders=false
+io.tracee.contextlogger.servlet.contextprovider.ServletRequestContextProvider.httpRequestAttributes=false
+io.tracee.contextlogger.servlet.contextprovider.ServletRequestContextProvider.cookies=false
+io.tracee.contextlogger.servlet.contextprovider.ServletRequestContextProvider.httpRemoteAddress=false
+io.tracee.contextlogger.servlet.contextprovider.ServletRequestContextProvider.httpRemoteHost=false
+io.tracee.contextlogger.servlet.contextprovider.ServletRequestContextProvider.httpRemotePort=false
+io.tracee.contextlogger.servlet.contextprovider.ServletRequestContextProvider.scheme=false
+io.tracee.contextlogger.servlet.contextprovider.ServletRequestContextProvider.secure=false
+io.tracee.contextlogger.servlet.contextprovider.ServletRequestContextProvider.contentType=false
+io.tracee.contextlogger.servlet.contextprovider.ServletRequestContextProvider.contentLength=false
+io.tracee.contextlogger.servlet.contextprovider.ServletRequestContextProvider.locale=false
 
 # servlet response
-io.tracee.contextlogger.contextprovider.servlet.ServletResponseContextProvider.httpStatusCode=true
-io.tracee.contextlogger.contextprovider.servlet.ServletResponseContextProvider.httpResponseHeaders=false
+io.tracee.contextlogger.servlet.contextprovider.ServletResponseContextProvider.httpStatusCode=true
+io.tracee.contextlogger.servlet.contextprovider.ServletResponseContextProvider.httpResponseHeaders=false
 
 # servlet session
-io.tracee.contextlogger.contextprovider.servlet.ServletSessionContextProvider=false
+io.tracee.contextlogger.servlet.contextprovider.ServletSessionContextProvider=false
 
 # javaee
-io.tracee.contextlogger.contextprovider.javaee.InvocationContextContextProvider.typeName=true
-io.tracee.contextlogger.contextprovider.javaee.InvocationContextContextProvider.methodName=true
-io.tracee.contextlogger.contextprovider.javaee.InvocationContextContextProvider.parameters=true
-io.tracee.contextlogger.contextprovider.javaee.InvocationContextContextProvider.serializedTargetInstance=false
+io.tracee.contextlogger.javaee.contextprovider.InvocationContextContextProvider.typeName=true
+io.tracee.contextlogger.javaee.contextprovider.InvocationContextContextProvider.methodName=true
+io.tracee.contextlogger.javaee.contextprovider.InvocationContextContextProvider.parameters=true
+io.tracee.contextlogger.javaee.contextprovider.InvocationContextContextProvider.serializedTargetInstance=false
 
 # jaxws
-io.tracee.contextlogger.contextprovider.jaxws.JaxWsContextProvider.soapRequest=true
-io.tracee.contextlogger.contextprovider.jaxws.JaxWsContextProvider.soapResponse=true
+io.tracee.contextlogger.jaxws.contextprovider.JaxWsContextProvider.soapRequest=true
+io.tracee.contextlogger.jaxws.contextprovider.JaxWsContextProvider.soapResponse=true
