+++ impl/src/main/resources/TraceeContextLoggerFullProfile.properties
@@ -26,47 +26,47 @@ io.tracee.contextlogger.contextprovider.java.JavaThrowableContextProvider.messag
 io.tracee.contextlogger.contextprovider.java.JavaThrowableContextProvider.stacktrace=true
 
 # Watchdog context data
-io.tracee.contextlogger.contextprovider.tracee.WatchdogContextProvider=true
-io.tracee.contextlogger.contextprovider.tracee.WatchdogContextProvider.id=true
-io.tracee.contextlogger.contextprovider.tracee.WatchdogContextProvider.proceedingJoinPoint=true
+io.tracee.contextlogger.watchdog.contextprovider.WatchdogContextProvider=true
+io.tracee.contextlogger.watchdog.contextprovider.WatchdogContextProvider.id=true
+io.tracee.contextlogger.watchdog.contextprovider.WatchdogContextProvider.proceedingJoinPoint=true
 
 # ASPECTJ - ProceedingJoinPoint
-io.tracee.contextlogger.contextprovider.aspectj.AspectjProceedingJoinPointContextProvider=true
-io.tracee.contextlogger.contextprovider.aspectj.AspectjProceedingJoinPointContextProvider.clazz=true
-io.tracee.contextlogger.contextprovider.aspectj.AspectjProceedingJoinPointContextProvider.method=true
-io.tracee.contextlogger.contextprovider.aspectj.AspectjProceedingJoinPointContextProvider.parameters=true
-io.tracee.contextlogger.contextprovider.aspectj.AspectjProceedingJoinPointContextProvider.serializedTargetInstance=true
+io.tracee.contextlogger.watchdog.contextprovider.AspectjProceedingJoinPointContextProvider=true
+io.tracee.contextlogger.watchdog.contextprovider.AspectjProceedingJoinPointContextProvider.clazz=true
+io.tracee.contextlogger.watchdog.contextprovider.AspectjProceedingJoinPointContextProvider.method=true
+io.tracee.contextlogger.watchdog.contextprovider.AspectjProceedingJoinPointContextProvider.parameters=true
+io.tracee.contextlogger.watchdog.contextprovider.AspectjProceedingJoinPointContextProvider.serializedTargetInstance=true
 
 # servlet request
-io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider=true
-io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider.url=true
-io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider.httpMethod=true
-io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider.httpParameters=true
-io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider.httpRequestHeaders=true
-io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider.httpRequestAttributes=true
-io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider.cookies=true
-io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider.httpRemoteAddress=true
-io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider.httpRemoteHost=true
-io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider.httpRemotePort=true
-io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider.scheme=true
-io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider.secure=true
-io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider.contentType=true
-io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider.contentLength=true
-io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider.locale=true
+io.tracee.contextlogger.servlet.contextprovider.ServletRequestContextProvider=true
+io.tracee.contextlogger.servlet.contextprovider.ServletRequestContextProvider.url=true
+io.tracee.contextlogger.servlet.contextprovider.ServletRequestContextProvider.httpMethod=true
+io.tracee.contextlogger.servlet.contextprovider.ServletRequestContextProvider.httpParameters=true
+io.tracee.contextlogger.servlet.contextprovider.ServletRequestContextProvider.httpRequestHeaders=true
+io.tracee.contextlogger.servlet.contextprovider.ServletRequestContextProvider.httpRequestAttributes=true
+io.tracee.contextlogger.servlet.contextprovider.ServletRequestContextProvider.cookies=true
+io.tracee.contextlogger.servlet.contextprovider.ServletRequestContextProvider.httpRemoteAddress=true
+io.tracee.contextlogger.servlet.contextprovider.ServletRequestContextProvider.httpRemoteHost=true
+io.tracee.contextlogger.servlet.contextprovider.ServletRequestContextProvider.httpRemotePort=true
+io.tracee.contextlogger.servlet.contextprovider.ServletRequestContextProvider.scheme=true
+io.tracee.contextlogger.servlet.contextprovider.ServletRequestContextProvider.secure=true
+io.tracee.contextlogger.servlet.contextprovider.ServletRequestContextProvider.contentType=true
+io.tracee.contextlogger.servlet.contextprovider.ServletRequestContextProvider.contentLength=true
+io.tracee.contextlogger.servlet.contextprovider.ServletRequestContextProvider.locale=true
 
 # servlet response
-io.tracee.contextlogger.contextprovider.servlet.ServletResponseContextProvider.httpStatusCode=true
-io.tracee.contextlogger.contextprovider.servlet.ServletResponseContextProvider.httpResponseHeaders=true
+io.tracee.contextlogger.servlet.contextprovider.ServletResponseContextProvider.httpStatusCode=true
+io.tracee.contextlogger.servlet.contextprovider.ServletResponseContextProvider.httpResponseHeaders=true
 
 # servlet session
-io.tracee.contextlogger.contextprovider.servlet.ServletSessionContextProvider=true
+io.tracee.contextlogger.servlet.contextprovider.ServletSessionContextProvider=true
 
 # javaee
-io.tracee.contextlogger.contextprovider.javaee.InvocationContextContextProvider.typeName=true
-io.tracee.contextlogger.contextprovider.javaee.InvocationContextContextProvider.methodName=true
-io.tracee.contextlogger.contextprovider.javaee.InvocationContextContextProvider.parameters=true
-io.tracee.contextlogger.contextprovider.javaee.InvocationContextContextProvider.serializedTargetInstance=true
+io.tracee.contextlogger.javaee.contextprovider.InvocationContextContextProvider.typeName=true
+io.tracee.contextlogger.javaee.contextprovider.InvocationContextContextProvider.methodName=true
+io.tracee.contextlogger.javaee.contextprovider.InvocationContextContextProvider.parameters=true
+io.tracee.contextlogger.javaee.contextprovider.InvocationContextContextProvider.serializedTargetInstance=true
 
 # jaxws
-io.tracee.contextlogger.contextprovider.jaxws.JaxWsContextProvider.soapRequest=true
-io.tracee.contextlogger.contextprovider.jaxws.JaxWsContextProvider.soapResponse=true
+io.tracee.contextlogger.jaxws.contextprovider.JaxWsContextProvider.soapRequest=true
+io.tracee.contextlogger.jaxws.contextprovider.JaxWsContextProvider.soapResponse=true
