+++ impl/src/main/resources/io.tracee.contextlogger.internal.wrappercontextproviders
@@ -1,14 +1,14 @@
-io.tracee.contextlogger.contextprovider.aspectj.AspectjProceedingJoinPointContextProvider
-io.tracee.contextlogger.contextprovider.javaee.InvocationContextContextProvider
+io.tracee.contextlogger.watchdog.contextprovider.AspectjProceedingJoinPointContextProvider
+io.tracee.contextlogger.javaee.contextprovider.InvocationContextContextProvider
 io.tracee.contextlogger.contextprovider.java.JavaThrowableContextProvider
-io.tracee.contextlogger.contextprovider.jaxws.JaxWsContextProvider
-io.tracee.contextlogger.contextprovider.servlet.ServletCookieContextProvider
-io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider
-io.tracee.contextlogger.contextprovider.servlet.ServletResponseContextProvider
-io.tracee.contextlogger.contextprovider.servlet.ServletSessionContextProvider
+io.tracee.contextlogger.jaxws.contextprovider.JaxWsContextProvider
+io.tracee.contextlogger.servlet.contextprovider.ServletCookieContextProvider
+io.tracee.contextlogger.servlet.contextprovider.ServletRequestContextProvider
+io.tracee.contextlogger.servlet.contextprovider.ServletResponseContextProvider
+io.tracee.contextlogger.servlet.contextprovider.ServletSessionContextProvider
 io.tracee.contextlogger.contextprovider.tracee.TraceeMessageContextProvider
 io.tracee.contextlogger.contextprovider.tracee.PassedDataContextProvider
-io.tracee.contextlogger.contextprovider.tracee.WatchdogContextProvider
+io.tracee.contextlogger.watchdog.contextprovider.WatchdogContextProvider
 io.tracee.contextlogger.contextprovider.utility.NameValuePair
