+++ impl/src/test/java/io/tracee/contextlogger/impl/ContextLoggerConfigurationTest.java
@@ -1,17 +1,15 @@
 package io.tracee.contextlogger.impl;
 
-import io.tracee.contextlogger.api.ImplicitContext;
-import io.tracee.contextlogger.contextprovider.java.JavaThrowableContextProvider;
-import io.tracee.contextlogger.contextprovider.servlet.ServletRequestContextProvider;
-import io.tracee.contextlogger.profile.Profile;
-import io.tracee.contextlogger.profile.ProfilePropertyNames;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.junit.Before;
 import org.junit.Ignore;
 import org.junit.Test;
 
-import javax.servlet.http.HttpServletRequest;
+import io.tracee.contextlogger.api.ImplicitContext;
+import io.tracee.contextlogger.contextprovider.java.JavaThrowableContextProvider;
+import io.tracee.contextlogger.profile.Profile;
+import io.tracee.contextlogger.profile.ProfilePropertyNames;
 
 /**
  * Test class for {@link ContextLoggerConfiguration}.
@@ -23,7 +21,6 @@ public class ContextLoggerConfigurationTest {
     public void init() {
         System.setProperty(ProfilePropertyNames.PROFILE_SET_GLOBALLY_VIA_SYSTEM_PROPERTIES, Profile.BASIC.name());
     }
-
 
     @Test
     public void should_get_context_logger_configuration() {
@@ -53,11 +50,11 @@ public class ContextLoggerConfigurationTest {
 
         MatcherAssert.assertThat(contextLoggerConfiguration, Matchers.notNullValue());
         MatcherAssert.assertThat(contextLoggerConfiguration.getClassToWrapperMap(), Matchers.notNullValue());
-        MatcherAssert.assertThat(contextLoggerConfiguration.getClassToWrapperMap().containsKey(HttpServletRequest.class), Matchers.equalTo(true));
-        MatcherAssert.assertThat(contextLoggerConfiguration.getClassToWrapperMap().get(HttpServletRequest.class).equals(ServletRequestContextProvider.class), Matchers.equalTo(true));
+        MatcherAssert.assertThat(contextLoggerConfiguration.getClassToWrapperMap().containsKey(Throwable.class), Matchers.equalTo(true));
+        MatcherAssert.assertThat(contextLoggerConfiguration.getClassToWrapperMap().get(Throwable.class).equals(JavaThrowableContextProvider.class),
+                Matchers.equalTo(true));
 
     }
-
 
     @Test
     @Ignore
@@ -71,7 +68,6 @@ public class ContextLoggerConfigurationTest {
 
     }
 
-
     @Test
     public void should_get_wrapper_classes_of_context_logger_configuration() {
 
@@ -79,11 +75,9 @@ public class ContextLoggerConfigurationTest {
 
         MatcherAssert.assertThat(contextLoggerConfiguration, Matchers.notNullValue());
         MatcherAssert.assertThat(contextLoggerConfiguration.getWrapperClasses(), Matchers.notNullValue());
-        MatcherAssert.assertThat(contextLoggerConfiguration.getWrapperClasses().contains(ServletRequestContextProvider.class), Matchers.equalTo(true));
         MatcherAssert.assertThat(contextLoggerConfiguration.getWrapperClasses().contains(JavaThrowableContextProvider.class), Matchers.equalTo(true));
 
     }
-
 
     @Test
     public void should_get_wrappers_of_context_logger_configuration() {
