+++ jaxws/src/main/java/io/tracee/contextlogger/jaxws/container/AbstractTraceeErrorLoggingHandler.java
@@ -16,7 +16,7 @@ import io.tracee.contextlogger.MessagePrefixProvider;
 import io.tracee.contextlogger.TraceeContextLogger;
 import io.tracee.contextlogger.api.ImplicitContext;
 import io.tracee.contextlogger.api.internal.MessageLogLevel;
-import io.tracee.contextlogger.contextprovider.jaxws.JaxWsWrapper;
+import io.tracee.contextlogger.jaxws.contextprovider.JaxWsWrapper;
 import io.tracee.jaxws.AbstractTraceeHandler;
 import io.tracee.jaxws.container.TraceeServerHandler;
