+++ jaxws/src/main/java/io/tracee/contextlogger/jaxws/contextprovider/JaxWsWrapper.java
@@ -1,0 +1,30 @@
+package io.tracee.contextlogger.jaxws.contextprovider;
+
+
+/**
+ * Wrapper class for JaxWs soap request and response.
+ */
+public final class JaxWsWrapper {
+
+    private final String soapRequest;
+    private final String soapResponse;
+
+    private JaxWsWrapper(final String soapRequest, final String soapResponse) {
+        this.soapRequest = soapRequest;
+        this.soapResponse = soapResponse;
+
+    }
+
+    public static JaxWsWrapper wrap(final String soapRequest, final String soapResponse) {
+        return new JaxWsWrapper(soapRequest, soapResponse);
+    }
+
+    public String getSoapRequest() {
+        return soapRequest;
+    }
+
+    public String getSoapResponse() {
+        return soapResponse;
+    }
+
+}
