+++ servlet/pom.xml
@@ -21,10 +21,12 @@
         </dependency>
 
 		<!-- provided external dependencies-->
-        <dependency>
-			<groupId>javax.servlet</groupId>
-			<artifactId>servlet-api</artifactId>
-        </dependency>
+		<dependency>
+			<groupId>org.jboss.spec.javax.servlet</groupId>
+			<artifactId>jboss-servlet-api_3.0_spec</artifactId>
+			<version>1.0.0.Final</version>
+			<scope>provided</scope>
+		</dependency>
 
 		<!-- unit test dependencies -->
 		<dependency>
